requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
Django==2.2.2
idna==3.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
